perf(blueprints): server-render the description markdown; fix Shiki on workerd
Extends the SSR markdown work to blueprint pages, where the whole detail
island is client:only (the vue-flow topology can't SSR) and the
description was therefore invisible to crawlers and popped in after
hydration — the largest CLS contributor on blueprint pages (0.483 in CI).
The description now renders as its own server-rendered island
(BlueprintMarkdown > Suspense > async BlueprintMarkdownBody, the same
shape as PluginsMDCRender on plugin pages).

Doing so surfaced a pre-existing production bug: getMarked() used the
full `shiki` bundle, whose default Oniguruma WASM engine workerd refuses
to compile ("Wasm code generation disallowed by embedder"), so every
fenced code block in server-rendered markdown silently rendered empty.
getMarked() now uses the shared core highlighter (JavaScript regex
engine, same as the schema code blocks), which also drops the
all-languages shiki build from the client MDC chunk. bash and json
grammars are added to the shared toolset; unregistered languages fall
back to plain text.
